Fix typo in dictionary_reader.py.
We check for the absence of a lemma using "if lemma is None". But the
variable would never be None because it's initialized to the empty
string. It should be initialized to None in order for the check to make
sense.

In order to avoid crashing the dictionary reader, we log an error
instead of raising an exception.